+++ app/src/js/components/Granules/list.js
@@ -29,7 +29,6 @@ import { strings } from '../locale';
 import { workflowOptionNames } from '../../selectors';
 import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
 import ListFilters from '../ListActions/ListFilters';
-import pageSizeOptions from '../../utils/page-size';
 
 const AllGranules = ({
   collections,
@@ -143,6 +142,8 @@ const AllGranules = ({
           bulkActions={generateBulkActions()}
           rowId="granuleId"
           sortId={tablesortId}
+          filterAction={filterGranules}
+          filterClear={clearGranulesFilter}
         >
           <ListFilters>
             <Dropdown
@@ -174,16 +175,6 @@ const AllGranules = ({
               clear={clearGranulesSearch}
               label="Search"
               placeholder="Granule ID"
-            />
-            <Dropdown
-              options={pageSizeOptions}
-              action={filterGranules}
-              clear={clearGranulesFilter}
-              paramKey="limit"
-              label="Results Per Page"
-              inputProps={{
-                placeholder: 'Results Per Page',
-              }}
             />
           </ListFilters>
         </List>