+++ app/src/js/components/Granules/overview.js
@@ -35,7 +35,6 @@ import { workflowOptionNames } from '../../selectors';
 import { window } from '../../utils/browser';
 import ListFilters from '../ListActions/ListFilters';
 import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
-import pageSizeOptions from '../../utils/page-size';
 import { downloadFile } from '../../utils/download-file';
 
 const breadcrumbConfig = [
@@ -174,6 +173,8 @@ class GranulesOverview extends React.Component {
             bulkActions={this.generateBulkActions()}
             rowId='granuleId'
             sortId='timestamp'
+            filterAction={filterGranules}
+            filterClear={clearGranulesFilter}
           >
             <ListFilters>
               <Dropdown
@@ -203,16 +204,6 @@ class GranulesOverview extends React.Component {
                 clear={clearGranulesSearch}
                 label='Search'
                 placeholder='Granule ID'
-              />
-              <Dropdown
-                options={pageSizeOptions}
-                action={filterGranules}
-                clear={clearGranulesFilter}
-                paramKey='limit'
-                label='Results Per Page'
-                inputProps={{
-                  placeholder: 'Results Per Page'
-                }}
               />
             </ListFilters>
           </List>