+++ CHANGELOG.md
@@ -12,12 +12,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Add some ESLint configuration to improve code quality
   - Upgrades Cypress to 4.12.0
 
+- **CUMULUS-1892**
+  - Added TableHeader component to allow user to select page and limit params
+  - Removed all Results Per Page dropdowns in favor of the header
+
 ### Fixed
 
 - **CUMULUS-1876**
   - Fix/remove unnecessary timers on home page
 
-### Fixed
 - **CUMULUS-2121**
   - Fix PDR list page with status filter
   - Update individual PDR page to properly display granules